backend/foodgramm/requirements.txt
Django==2.2.19
psycopg2-binary==2.8.6
python-dotenv==0.21.1
Pillow==9.2.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-filter==21.1
drf-extra-fields==3.2.1
fpdf==1.7.2
flake8==5.0.4
gunicorn==20.0.4